+++ component/componenttest/nop_processor.go
@@ -21,7 +21,6 @@ import (
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.opentelemetry.io/collector/internal/internalinterface"
 )
 
 // NewNopProcessorCreateSettings returns a new nop settings for Create*Processor functions.
@@ -36,57 +35,32 @@ type nopProcessorConfig struct {
 	config.ProcessorSettings `mapstructure:",squash"` // squash ensures fields are correctly decoded in embedded struct
 }
 
-// nopProcessorFactory is factory for nopProcessor.
-type nopProcessorFactory struct {
-	internalinterface.BaseInternal
-}
-
-var nopProcessorFactoryInstance = &nopProcessorFactory{}
+var nopProcessorFactory = component.NewProcessorFactory(
+	"nop",
+	func() config.Processor {
+		return &nopProcessorConfig{
+			ProcessorSettings: config.NewProcessorSettings(config.NewComponentID("nop")),
+		}
+	},
+	component.WithTracesProcessor(createTracesProcessor),
+	component.WithMetricsProcessor(createMetricsProcessor),
+	component.WithLogsProcessor(createLogsProcessor),
+)
 
 // NewNopProcessorFactory returns a component.ProcessorFactory that constructs nop processors.
 func NewNopProcessorFactory() component.ProcessorFactory {
-	return nopProcessorFactoryInstance
+	return nopProcessorFactory
 }
 
-// Type gets the type of the Processor config created by this factory.
-func (f *nopProcessorFactory) Type() config.Type {
-	return "nop"
-}
-
-// CreateDefaultConfig creates the default configuration for the Processor.
-func (f *nopProcessorFactory) CreateDefaultConfig() config.Processor {
-	return &nopProcessorConfig{
-		ProcessorSettings: config.NewProcessorSettings(config.NewComponentID("nop")),
-	}
-}
-
-// CreateTracesProcessor implements component.ProcessorFactory interface.
-func (f *nopProcessorFactory) CreateTracesProcessor(
-	_ context.Context,
-	_ component.ProcessorCreateSettings,
-	_ config.Processor,
-	_ consumer.Traces,
-) (component.TracesProcessor, error) {
+func createTracesProcessor(context.Context, component.ProcessorCreateSettings, config.Processor, consumer.Traces) (component.TracesProcessor, error) {
 	return nopProcessorInstance, nil
 }
 
-// CreateMetricsProcessor implements component.ProcessorFactory interface.
-func (f *nopProcessorFactory) CreateMetricsProcessor(
-	_ context.Context,
-	_ component.ProcessorCreateSettings,
-	_ config.Processor,
-	_ consumer.Metrics,
-) (component.MetricsProcessor, error) {
+func createMetricsProcessor(context.Context, component.ProcessorCreateSettings, config.Processor, consumer.Metrics) (component.MetricsProcessor, error) {
 	return nopProcessorInstance, nil
 }
 
-// CreateLogsProcessor implements component.ProcessorFactory interface.
-func (f *nopProcessorFactory) CreateLogsProcessor(
-	_ context.Context,
-	_ component.ProcessorCreateSettings,
-	_ config.Processor,
-	_ consumer.Logs,
-) (component.LogsProcessor, error) {
+func createLogsProcessor(context.Context, component.ProcessorCreateSettings, config.Processor, consumer.Logs) (component.LogsProcessor, error) {
 	return nopProcessorInstance, nil
 }