+++ receiver/receiverhelper/factory.go
@@ -15,118 +15,32 @@
 package receiverhelper // import "go.opentelemetry.io/collector/receiver/receiverhelper"
 
 import (
-	"context"
-
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/component/componenterror"
-	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/internal/internalinterface"
 )
 
-// FactoryOption apply changes to ReceiverOptions.
-type FactoryOption func(o *factory)
+// Deprecated: [v0.46.0] use component.ReceiverFactoryOption.
+type FactoryOption = component.ReceiverFactoryOption
 
-// WithTraces overrides the default "error not supported" implementation for CreateTracesReceiver.
-func WithTraces(createTracesReceiver CreateTracesReceiver) FactoryOption {
-	return func(o *factory) {
-		o.createTracesReceiver = createTracesReceiver
-	}
-}
+// Deprecated: [v0.46.0] use component.WithTracesReceiver.
+var WithTraces = component.WithTracesReceiver
 
-// WithMetrics overrides the default "error not supported" implementation for CreateMetricsReceiver.
-func WithMetrics(createMetricsReceiver CreateMetricsReceiver) FactoryOption {
-	return func(o *factory) {
-		o.createMetricsReceiver = createMetricsReceiver
-	}
-}
+// Deprecated: [v0.46.0] use component.WithMetricsReceiver.
+var WithMetrics = component.WithMetricsReceiver
 
-// WithLogs overrides the default "error not supported" implementation for CreateLogsReceiver.
-func WithLogs(createLogsReceiver CreateLogsReceiver) FactoryOption {
-	return func(o *factory) {
-		o.createLogsReceiver = createLogsReceiver
-	}
-}
+// Deprecated: [v0.46.0] use component.WithLogsReceiver.
+var WithLogs = component.WithLogsReceiver
 
-// CreateDefaultConfig is the equivalent of component.ReceiverFactory.CreateDefaultConfig()
-type CreateDefaultConfig func() config.Receiver
+// Deprecated: [v0.46.0] use component.ReceiverCreateDefaultConfigFunc.
+type CreateDefaultConfig = component.ReceiverCreateDefaultConfigFunc
 
-// CreateTracesReceiver is the equivalent of component.ReceiverFactory.CreateTracesReceiver()
-type CreateTracesReceiver func(context.Context, component.ReceiverCreateSettings, config.Receiver, consumer.Traces) (component.TracesReceiver, error)
+// Deprecated: [v0.46.0] use component.CreateTracesReceiverFunc.
+type CreateTracesReceiver = component.CreateTracesReceiverFunc
 
-// CreateMetricsReceiver is the equivalent of component.ReceiverFactory.CreateMetricsReceiver()
-type CreateMetricsReceiver func(context.Context, component.ReceiverCreateSettings, config.Receiver, consumer.Metrics) (component.MetricsReceiver, error)
+// Deprecated: [v0.46.0] use component.CreateMetricsReceiverFunc.
+type CreateMetricsReceiver = component.CreateMetricsReceiverFunc
 
-// CreateLogsReceiver is the equivalent of component.ReceiverFactory.CreateLogsReceiver()
-type CreateLogsReceiver func(context.Context, component.ReceiverCreateSettings, config.Receiver, consumer.Logs) (component.LogsReceiver, error)
+// Deprecated: [v0.46.0] use component.CreateLogsReceiverFunc.
+type CreateLogsReceiver = component.CreateLogsReceiverFunc
 
-type factory struct {
-	internalinterface.BaseInternal
-	cfgType               config.Type
-	createDefaultConfig   CreateDefaultConfig
-	createTracesReceiver  CreateTracesReceiver
-	createMetricsReceiver CreateMetricsReceiver
-	createLogsReceiver    CreateLogsReceiver
-}
-
-// NewFactory returns a component.ReceiverFactory.
-func NewFactory(
-	cfgType config.Type,
-	createDefaultConfig CreateDefaultConfig,
-	options ...FactoryOption) component.ReceiverFactory {
-	f := &factory{
-		cfgType:             cfgType,
-		createDefaultConfig: createDefaultConfig,
-	}
-	for _, opt := range options {
-		opt(f)
-	}
-	return f
-}
-
-// Type gets the type of the Receiver config created by this factory.
-func (f *factory) Type() config.Type {
-	return f.cfgType
-}
-
-// CreateDefaultConfig creates the default configuration for receiver.
-func (f *factory) CreateDefaultConfig() config.Receiver {
-	return f.createDefaultConfig()
-}
-
-// CreateTracesReceiver creates a component.TracesReceiver based on this config.
-func (f *factory) CreateTracesReceiver(
-	ctx context.Context,
-	set component.ReceiverCreateSettings,
-	cfg config.Receiver,
-	nextConsumer consumer.Traces) (component.TracesReceiver, error) {
-	if f.createTracesReceiver != nil {
-		return f.createTracesReceiver(ctx, set, cfg, nextConsumer)
-	}
-	return nil, componenterror.ErrDataTypeIsNotSupported
-}
-
-// CreateMetricsReceiver creates a component.MetricsReceiver based on this config.
-func (f *factory) CreateMetricsReceiver(
-	ctx context.Context,
-	set component.ReceiverCreateSettings,
-	cfg config.Receiver,
-	nextConsumer consumer.Metrics) (component.MetricsReceiver, error) {
-	if f.createMetricsReceiver != nil {
-		return f.createMetricsReceiver(ctx, set, cfg, nextConsumer)
-	}
-	return nil, componenterror.ErrDataTypeIsNotSupported
-}
-
-// CreateLogsReceiver creates a component.LogsReceiver based on this config.
-func (f *factory) CreateLogsReceiver(
-	ctx context.Context,
-	set component.ReceiverCreateSettings,
-	cfg config.Receiver,
-	nextConsumer consumer.Logs,
-) (component.LogsReceiver, error) {
-	if f.createLogsReceiver != nil {
-		return f.createLogsReceiver(ctx, set, cfg, nextConsumer)
-	}
-	return nil, componenterror.ErrDataTypeIsNotSupported
-}
+// Deprecated: [v0.46.0] use component.NewReceiverFactory.
+var NewFactory = component.NewReceiverFactory